Move snx utilities into own file.
